% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{feffect}
\alias{feffect}
\title{felling effect}
\usage{
feffect(day, decay = 1, b1 = 100, a1 = 0.25, type = 2)
}
\arguments{
\item{day}{The number of simulations to use to evaluate confidence intervals.}

\item{decay}{A number.}

\item{b1}{A number.}

\item{a1}{A number.}

\item{type}{A number.}
}
\description{
This function does everything from read in files, select calibtation period and
thens runs a calibration and saves the file.
}
\examples{
1 + 1
}

