% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{fitDecay}
\alias{fitDecay}
\title{felling effect}
\usage{
fitDecay(decay, gcv = FALSE, data)
}
\arguments{
\item{decay}{The number of simulations to use to evaluate confidence intervals.}

\item{gcv}{A number.}

\item{data}{A number.}
}
\description{
This function does everything from read in files, select calibtation period and
thens runs a calibration and saves the file.
}
\examples{
1 + 1
}

