% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{getfd}
\alias{getfd}
\title{Run a full calibration}
\usage{
getfd(object, nbsplines = 24, lambda = 0, breaks = NULL)
}
\arguments{
\item{object}{The number of simulations to use to evaluate confidence intervals.}

\item{nbsplines}{A number.}

\item{lambda}{A number.}

\item{breaks}{A number.}
}
\description{
This function does everything from read in files, select calibtation period and
thens runs a calibration and saves the file.
}
\examples{
1 + 1
}

