% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{getFelledPC}
\alias{getFelledPC}
\title{felling effect}
\usage{
getFelledPC(model, sim = FALSE, n = 1, which = grepl("fell",
  names(coef(model))))
}
\arguments{
\item{model}{The number of simulations to use to evaluate confidence intervals.}

\item{sim}{A number.}

\item{n}{A number.}

\item{which}{A number.}
}
\description{
This function does everything from read in files, select calibtation period and
thens runs a calibration and saves the file.
}
\examples{
1 + 1
}

