% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_pcs}
\alias{plot_pcs}
\title{Run a full calibration}
\usage{
plot_pcs(infodat, main = NULL)
}
\arguments{
\item{infodat}{The number of simulations to use to evaluate confidence intervals.}

\item{main}{A number.}
}
\description{
This function does everything from read in files, select calibtation period and
thens runs a calibration and saves the file.
}
\examples{
1 + 1
}

