% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotPCbyblock}
\alias{plotPCbyblock}
\title{Run a full calibration}
\usage{
plotPCbyblock(dat, PC = "PC1", ncuts = 4)
}
\arguments{
\item{dat}{The number of simulations to use to evaluate confidence intervals.}

\item{PC}{A number.}

\item{ncuts}{A number.}
}
\description{
This function does everything from read in files, select calibtation period and
thens runs a calibration and saves the file.
}
\examples{
1 + 1
}

