% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{reduceData}
\alias{reduceData}
\title{Run a full calibration}
\usage{
reduceData(dat, basisdim = 47, redbasisdim = 12, nharm = 4)
}
\arguments{
\item{dat}{The number of simulations to use to evaluate confidence intervals.}

\item{basisdim}{The number of simulations to use to evaluate confidence intervals.}

\item{redbasisdim}{The number of simulations to use to evaluate confidence intervals.}

\item{nharm}{The number of simulations to use to evaluate confidence intervals.}
}
\description{
This function does everything from read in files, select calibtation period and
thens runs a calibration and saves the file.
}
\examples{
1 + 1
}

