% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{reduceData_old}
\alias{reduceData_old}
\title{Run a full calibration}
\usage{
reduceData_old(object, pc)
}
\arguments{
\item{object}{The number of simulations to use to evaluate confidence intervals.}

\item{pc}{A number.}
}
\description{
This function does everything from read in files, select calibtation period and
thens runs a calibration and saves the file.
}
\examples{
1 + 1
}

