% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lochard-data.R
\docType{data}
\name{stream}
\alias{stream}
\title{Raw temperature data.}
\format{a \code{data.frame}.}
\source{
Marine Scotland Science, Freshwater Laboratory
}
\usage{
stream
}
\description{
raw water temperature readings from 5 rivers in the loch ard catchment.
}
\author{
Iain Malcolm, Karen Millidine, Marine Scotland Science, Crown Copywrite 2015
}
\keyword{datasets}

